requirements.txt
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
dnspython==2.2.0
email-validator==1.1.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Limiter==2.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
limits==2.3.1
MarkupSafe==2.0.1
numpy==1.22.1
pandas==1.3.5
pycparser==2.21
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.30
Werkzeug==2.0.2
WTForms==3.0.1
pycryptodome==3.13.0